+++ packages/wiz-ui-react/src/components/base/calendar/components/index.ts
@@ -1,0 +1,1 @@
+export { WizCalendar } from "./calendar";