+++ packages/wiz-ui-react/src/components/base/graphs/circle/components/types.ts
@@ -1,0 +1,4 @@
+export interface CircleGraphData {
+  label: string;
+  percentage: number;
+}