+++ packages/wiz-ui-react/src/components/base/graphs/line/stories/line-graph.stories.tsx
@@ -1,0 +1,17 @@
+import { Meta, StoryObj } from "@storybook/react";
+
+import { WizLineGraph } from "../components";
+
+const meta: Meta<typeof WizLineGraph> = {
+  title: "Base/Graphs/Line",
+  component: WizLineGraph,
+};
+
+export default meta;
+type Story = StoryObj<typeof WizLineGraph>;
+
+export const Default: Story = {
+  args: {
+    percentage: 56,
+  },
+};