+++ packages/wiz-ui-react/src/components/base/tables/card/components/card-td.tsx
@@ -1,0 +1,17 @@
+import { ComponentName } from "@wizleap-inc/wiz-ui-constants";
+import * as styles from "@wizleap-inc/wiz-ui-styles/bases/card-table.css";
+import { ComponentProps, ReactNode } from "react";
+
+type Props = {
+  children?: ReactNode;
+} & ComponentProps<"td">;
+
+const CardTd = ({ children, ...props }: Props) => (
+  <td className={styles.cardTdStyle} {...props}>
+    {children}
+  </td>
+);
+
+CardTd.displayName = ComponentName.CardTd;
+
+export const WizCardTd = CardTd;